+++ packages/playground/src/components/room/room-content.tsx
@@ -1,6 +1,6 @@
 import { MoveDownIcon, TriangleAlertIcon } from "lucide-react";
 import { observer } from "mobx-react-lite";
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import type { MCPToolResponse } from "@semoss/sdk";
 import {
 	Button,
@@ -19,8 +19,7 @@ import {
 	RoomInput,
 	RoomInputMenuPlugin,
 } from "@/components";
-import { LOADING_MESSAGES } from "@/constants";
-import { useAutoScroll } from "@/hooks";
+import { useAutoScroll, useLoadingMessage } from "@/hooks";
 import type { RoomStore } from "@/stores";
 
 interface RoomContentProps {
@@ -41,10 +40,7 @@ export const RoomContent: React.FC<RoomContentProps> = observer(({ room }) => {
 	const { setScrollEle, scrollToBottom, isUserScrolled } = useAutoScroll(
 		room.history?.length || 0,
 	);
-	/**
-	 * State
-	 */
-	const [loadingMessage, setLoadingMessage] = useState<string>("");
+	const loadingMessage = useLoadingMessage(room.isLoading);
 
 	/**
 	 * Functions
@@ -59,57 +55,9 @@ export const RoomContent: React.FC<RoomContentProps> = observer(({ room }) => {
 		return true;
 	};
 
-	console.log(
-		JSON.stringify({
-			loading: room.isLoading,
-			hasUnfinishedTools: room.hasUnfinishedTools,
-		}),
-	);
-
 	/**
 	 * Effects
 	 */
-
-	// iterate loading messages
-	useEffect(() => {
-		if (!room.isLoading) {
-			setLoadingMessage("");
-			return;
-		}
-
-		setLoadingMessage(LOADING_MESSAGES[0]);
-
-		let timeoutId: number | undefined;
-		let iteration = 1;
-		let cancelled = false;
-
-		const scheduleNext = () => {
-			const delay = 500 + 1500 * (iteration - 1);
-
-			timeoutId = window.setTimeout(() => {
-				if (cancelled) {
-					return;
-				}
-
-				const randomIndex =
-					1 +
-					Math.floor(Math.random() * (LOADING_MESSAGES.length - 1));
-
-				setLoadingMessage(LOADING_MESSAGES[randomIndex]);
-				iteration += 1;
-				scheduleNext();
-			}, delay);
-		};
-
-		scheduleNext();
-
-		return () => {
-			cancelled = true;
-			if (timeoutId !== undefined) {
-				window.clearTimeout(timeoutId);
-			}
-		};
-	}, [room.isLoading]);
 
 	// create a listener to process messages from the room
 	useEffect(() => {
