+++ packages/playground/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from "./useAutoScroll";
 export * from "./useChat";
 export * from "./useGlobalBreadcrumbs";
+export * from "./useLoadingMessage";
 export * from "./useRoot";
